+++ zenith-float-macro/src/lib.rs
@@ -192,6 +192,27 @@ fn root_fun(
     )))
 }
 
+fn bessel_j_fun(
+    expr: &ExprCall,
+    initial_err: usize,
+    err: &mut Vec<usize>,
+    cc: &mut Consts,
+) -> Result<TokenStream, Error> {
+    check_arg_num(2, expr)?;
+
+    let arg = traverse_expr(&expr.args[0], err, cc)?;
+    let n = &expr.args[1];
+    err.push(initial_err);
+
+    Ok(quote!(zenith_float::ExactNum::bessel_j(
+        &(#arg),
+        #n as usize,
+        p_wrk,
+        zenith_float::RoundingMode::None,
+        cc
+    )))
+}
+
 fn one_arg_fun_errcheck(
     fun: TokenStream,
     expr: &ExprCall,
@@ -282,7 +303,7 @@ fn traverse_call(
     err: &mut Vec<usize>,
     cc: &mut Consts,
 ) -> Result<TokenStream, Error> {
-    let errmes = "unexpected function name. Only \"recip\", \"sqrt\", \"cbrt\", \"root\", \"ln\", \"log2\", \"log10\", \"log\", \"log1p\", \"exp\", \"exp2\", \"exp10\", \"expm1\", \"pow\", \"rem_pi\", \"sin\", \"cos\", \"tan\", \"asin\", \"acos\", \"atan\", \"atan2\", \"hypot\", \"fma\", \"sinh\", \"cosh\", \"tanh\", \"asinh\", \"acosh\", \"atanh\" are allowed.";
+    let errmes = "unexpected function name. Only \"recip\", \"sqrt\", \"cbrt\", \"root\", \"ln\", \"log2\", \"log10\", \"log\", \"log1p\", \"exp\", \"exp2\", \"exp10\", \"expm1\", \"pow\", \"rem_pi\", \"sin\", \"cos\", \"tan\", \"asin\", \"acos\", \"atan\", \"atan2\", \"hypot\", \"fma\", \"mul_add\", \"sinh\", \"cosh\", \"tanh\", \"asinh\", \"acosh\", \"atanh\", \"erf\", \"erfc\", \"gamma\", \"ln_gamma\", \"bessel_j\" are allowed.";
 
     if let Expr::Path(fun) = expr.func.as_ref() {
         if let Some(fname) = fun.path.get_ident() {
@@ -464,6 +485,13 @@ fn traverse_call(
                     err,
                     cc,
                 ),
+                "mul_add" => three_arg_fun(
+                    quote!(zenith_float::ExactNum::mul_add),
+                    expr,
+                    2,
+                    err,
+                    cc,
+                ),
                 "sinh" => one_arg_fun(
                     quote!(zenith_float::ExactNum::sinh),
                     expr,
@@ -505,6 +533,11 @@ fn traverse_call(
                     quote!(zenith_float::macro_util::ErrAlgo::Atanh(&arg, emin)),
                     cc,
                 ),
+                "erf" => one_arg_fun(quote!(zenith_float::ExactNum::erf), expr, 2, err, cc, true),
+                "erfc" => one_arg_fun(quote!(zenith_float::ExactNum::erfc), expr, 2, err, cc, true),
+                "gamma" => one_arg_fun(quote!(zenith_float::ExactNum::gamma), expr, EXPONENT_BIT_SIZE + 1, err, cc, true),
+                "ln_gamma" => one_arg_fun(quote!(zenith_float::ExactNum::ln_gamma), expr, EXPONENT_BIT_SIZE + 1, err, cc, true),
+                "bessel_j" => bessel_j_fun(expr, 2, err, cc),
                 _ => return Err(Error::new(expr.span(), errmes)),
             }?;